Add changes from plugin Breadcrumbs.
